Quiz as modal on Welcome sign up page
