+++ client/src/App.jsx
@@ -8,6 +8,7 @@ import PrivateRoute from './components/PrivateRoute';
 import UserHomePage from './pages/UserHomePage';
 import Benefits from './pages/Benefits';
 import Quiz from './pages/Quiz/Quiz';
+import Welcome from './pages/Quiz/Welcome';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 import './App.css';
@@ -21,7 +22,7 @@ const App = () => {
           <Route exact path="/" component={AnonHomePage} />
           <Route exact path="/body-area" component={StretchCategory} />
           <Route exact path="/stretch/:id" component={IndividualStretch} />
-          <Route exact path="/sign-up" component={Quiz} />
+          <Route exact path="/sign-up" component={Welcome} />
           <PrivateRoute exact path="/profile" component={UserHomePage} />
         </Switch>
       </BrowserRouter>
